fast_neural_style/neural_style/requirements.txt
torch==0.2.0.post3
numpy==1.13.1
requests==2.18.4
torchvision==0.1.9
Pillow==4.2.1
